refactor: move ConcurrencyConflictError to platform, fix cast style, exit code 3, GUI handler
- Move ConcurrencyConflictError from system/_exceptions to platform/_exceptions
  and re-export from aignostics.platform — eliminates application→system
  bidirectional dependency
- Revert cast() string-form regression in runs.py (restore TC006-compliant
  cast("dict[str, Any]", ...) in all four call sites)
- CLI conflict handlers now exit 3 (distinct from exit 2 for not-found)
- Add specific ConcurrencyConflictError handler in GUI metadata editor
  with "reload and retry" guidance instead of generic error message
- Update all imports and test assertions accordingly
