+++ src/aignostics/platform/_exceptions.py
@@ -1,0 +1,10 @@
+"""Exceptions of platform module."""
+
+
+class ConcurrencyConflictError(ValueError):
+    """Raised when an optimistic concurrency precondition (HTTP 412) fails.
+
+    Subclasses ValueError so existing ``except ValueError`` callers still catch it,
+    while callers that need to distinguish a conflict from a bad-ID error can use
+    ``except ConcurrencyConflictError``.
+    """
