+++ src/aignostics/system/__init__.py
@@ -1,12 +1,10 @@
 """System module."""
 
 from ._cli import cli
-from ._exceptions import ConcurrencyConflictError
 from ._service import Service
 from ._settings import Settings
 
 __all__ = [
-    "ConcurrencyConflictError",
     "Service",
     "Settings",
     "cli",
