+++ tests/aignostics/application/service_test.py
@@ -8,8 +8,7 @@ import pytest
 from typer.testing import CliRunner
 
 from aignostics.application import Service as ApplicationService
-from aignostics.platform import ApiException, NotFoundException, RunData, RunOutput
-from aignostics.system import ConcurrencyConflictError
+from aignostics.platform import ApiException, ConcurrencyConflictError, NotFoundException, RunData, RunOutput
 from tests.constants_test import (
     HETA_APPLICATION_ID,
     HETA_APPLICATION_VERSION,
